+++ src/background/utils/constants.rs
@@ -6,7 +6,8 @@ lazy_static! {
         "Task Switching",
         "Task View",
         "Virtual desktop switching preview",
-        "Virtual desktop hotkey switching preview"
+        "Virtual desktop hotkey switching preview",
+        "Seelen Window Manager", // For some reason th WM is focused on change of virtual desktop
     ]
     .iter()
     .map(|x| x.to_string())